Added Vector header and cpp files.
Slight issue with cmath library and sqrt function where the machine returns one error stating that the magVector function has an undefined sqrt reference. It's weird and I don't know how to fix it because the solution looks exactly like what I'm doing. Maybe it's the nature of the file. Or maybe it's me.